app/requirements.txt
django==3.2
djangorestframework
django-filter
drf-yasg
psycopg2-binary
gunicorn
pyTelegramBotAPI